requirements.txt
streamlit>=1.20
pymupdf>=1.22.0
python-docx>=0.8.11
openai>=0.27.0
python-dotenv>=1.0.0
fpdf>=1.7.2
